requirements-devel.txt
pytest-django
ipdb
django-extensions
django-debug-toolbar
